feat(config,embed,search,storage): config cleanup, embed worker dedup, search docs and persistence, storage fixes
Config:
- Remove dead legacyLoadFromEnv; all env handling is in LoadDefaults/applyEnvVars/LoadFromFile
- Apply NORNICDB_EMBED_WORKER_NUM_WORKERS and NORNICDB_SEARCH_MIN_SIMILARITY in applyEnvVars
- Delete config.go.bak
- Add TestLoadFromEnv_EmbeddingWorkerNumWorkers

Embed worker (no double-processing):
- Serialize find+claim with claimMu so only one worker takes a node at a time
- Claim node (markNodeEmbedded) after validate, re-queue (AddToPendingEmbeddings) on embed/update failure
- Add AddToPendingEmbeddings on AsyncEngine, WALEngine, NamespacedEngine

Search:
- Document search index persistence, vector strategy, and HNSW tuning in configuration.md and searching.md
- Add search request logging (database, query) and 503 when index not ready (with Retry-After)
- Build indexes in background; IsSearchIndexReady for handler
- Remove search.go.bak

Storage:
- Allow int8/uint* in property validation (fix async flush _embedding_chunk_count gob decode)
- Store _embedding_chunk_count as int64 in badger_helpers

Misc:
- Clarify ChunkText has no per-node chunk limit (util/text_chunk.go, embed_queue test)

+++ pkg/storage/badger_helpers.go
@@ -201,7 +201,7 @@ func encodeNode(n *Node) ([]byte, bool, error) {
 		nodeCopy.Properties = make(map[string]any)
 	}
 	nodeCopy.Properties["_embeddings_stored_separately"] = true
-	nodeCopy.Properties["_embedding_chunk_count"] = len(embeddingsToStore)
+	nodeCopy.Properties["_embedding_chunk_count"] = int64(len(embeddingsToStore))
 
 	// Final encode with flag
 	data, err = encodeValue(&nodeCopy)
